refactor: init 함수 index.js로 이동, 함수명 initApp으로 변경
