+++ js/index.js
@@ -1,3 +1,16 @@
-import { init } from "./eventListener/init.js";
+import { handleCarNameInput } from "./controller/handleCarNameInput.js";
 
-init();
+const initEventListener = () => {
+  const carNameInput = document.querySelector("#carNameInput");
+  const carNameInputButton = document.querySelector(
+    "#carNameInputSubmitButton"
+  );
+  carNameInput.addEventListener("submit", handleCarNameInput);
+  carNameInputButton.addEventListener("click", handleCarNameInput);
+};
+
+export const initApp = () => {
+  initEventListener();
+};
+
+initApp();
